+++ src/app/shared/PersonSummary.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+  import cx from "classnames"
   import {modal} from "src/partials/state"
   import Anchor from "src/partials/Anchor.svelte"
   import PersonCircle from "src/app/shared/PersonCircle.svelte"
@@ -15,10 +16,12 @@
 </script>
 
 <div class="relative flex flex-grow flex-col gap-4 px-3 py-2">
-  <div class="flex items-start justify-between gap-4">
-    <Anchor on:click={inert ? null : showDetail} class="flex gap-4">
+  <div class="relative grid grid-cols-4 gap-4">
+    <Anchor
+      on:click={inert ? null : showDetail}
+      class={cx("col-span-3 flex gap-4 overflow-hidden", {"col-span-4": hideActions})}>
       <PersonCircle size={14} {pubkey} />
-      <div class="mr-16 flex flex-grow flex-col gap-1">
+      <div class="mr-16 flex min-w-0 flex-grow flex-col gap-1">
         <PersonName class="text-lg" {pubkey} />
         <PersonHandle {pubkey} />
       </div>