+++ src/app/views/OnboardingFollows.svelte
@@ -6,21 +6,25 @@
   import Content from "src/partials/Content.svelte"
   import PersonSummary from "src/app/shared/PersonSummary.svelte"
   import type {Profile} from "src/engine"
-  import {Env, Nip02, user, Directory, Builder} from "src/app/engine"
+  import {Directory, Builder} from "src/app/engine"
   import {modal} from "src/partials/state"
 
-  const {searchProfiles} = Directory
-  const follows = Nip02.graph
-    .key(user.getStateKey())
-    .derived(g => (g ? g.petnames.map(t => t[1]) : []))
+  export let petnames
 
-  if ($follows.length === 0) {
-    user.setPetnames(Env.DEFAULT_FOLLOWS.map(Builder.mention))
+  const {searchProfiles} = Directory
+
+  const addFollow = pubkey => {
+    petnames = petnames.concat([Builder.mention(pubkey)])
+  }
+
+  const removeFollow = pubkey => {
+    petnames = reject(t => t[1] === pubkey, petnames)
   }
 
   let q = ""
 
-  $: results = reject((p: Profile) => $follows.includes(p.pubkey), $searchProfiles(q))
+  $: pubkeys = petnames.map(t => t[1])
+  $: results = reject((p: Profile) => pubkeys.includes(p.pubkey), $searchProfiles(q))
 </script>
 
 <Content>
@@ -40,19 +44,19 @@
     <i class="fa fa-user-astronaut fa-lg" />
     <h2 class="staatliches text-2xl">Your follows</h2>
   </div>
-  {#if $follows.length === 0}
+  {#if pubkeys.length === 0}
     <div class="mt-8 flex items-center justify-center gap-2 text-center">
       <i class="fa fa-triangle-exclamation" />
       <span>No follows selected</span>
     </div>
   {:else}
-    {#each $follows as pubkey (pubkey)}
+    {#each pubkeys as pubkey (pubkey)}
       <PersonSummary {pubkey}>
         <div slot="actions">
           <Anchor
             theme="button"
             class="flex items-center gap-2"
-            on:click={() => user.unfollow(pubkey)}>
+            on:click={() => removeFollow(pubkey)}>
             <i class="fa fa-user-slash" /> Unfollow
           </Anchor>
         </div>
@@ -68,11 +72,11 @@
   </Input>
   {#each results.slice(0, 50) as profile (profile.pubkey)}
     <PersonSummary pubkey={profile.pubkey}>
-      <div slot="actions">
+      <div slot="actions" class="flex items-start justify-end">
         <Anchor
           theme="button-accent"
           class="flex items-center gap-2"
-          on:click={() => user.follow(profile.pubkey)}>
+          on:click={() => addFollow(profile.pubkey)}>
           <i class="fa fa-user-plus" /> Follow
         </Anchor>
       </div>