+++ src/app/views/OnboardingRelays.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-  import {pluck} from "ramda"
+  import type {Relay} from "src/engine"
+  import {pluck, reject, propEq} from "ramda"
   import {fuzzy} from "src/util/misc"
   import {modal} from "src/partials/state"
   import Input from "src/partials/Input.svelte"
@@ -7,22 +8,29 @@
   import Heading from "src/partials/Heading.svelte"
   import Content from "src/partials/Content.svelte"
   import RelayCard from "src/app/shared/RelayCard.svelte"
-  import {Nip65, user} from "src/app/engine"
+  import {Nip65} from "src/app/engine"
+
+  export let relays: Relay[]
 
   let q = ""
   let search
 
-  const userRelays = Nip65.policies.derived(() => user.getRelays())
-  const knownRelays = Nip65.relays
+  const {relays: known} = Nip65
 
-  $: {
-    const joined = new Set(pluck("url", $userRelays))
-
-    search = fuzzy(
-      $knownRelays.filter(r => !joined.has(r.url)),
-      {keys: ["name", "description", "url"]}
-    )
+  const addRelay = relay => {
+    relays = relays.concat(relay)
   }
+
+  const removeRelay = relay => {
+    relays = reject(propEq("url", relay.url), relays)
+  }
+
+  $: joined = new Set(pluck("url", relays))
+
+  $: search = fuzzy(
+    $known.filter(r => !joined.has(r.url)),
+    {keys: ["name", "description", "url"]}
+  )
 </script>
 
 <Content>
@@ -44,15 +52,25 @@
     <i class="fa fa-server fa-lg" />
     <h2 class="staatliches text-2xl">Your relays</h2>
   </div>
-  {#if $userRelays.length === 0}
+  {#if relays.length === 0}
     <div class="mt-8 flex items-center justify-center gap-2 text-center">
       <i class="fa fa-triangle-exclamation" />
       <span>No relays connected</span>
     </div>
   {:else}
     <div class="grid grid-cols-1 gap-4">
-      {#each $userRelays as relay (relay.url)}
-        <RelayCard {relay} />
+      {#each relays as relay (relay.url)}
+        <RelayCard {relay}>
+          <div slot="actions">
+            {#if relays.length > 1}
+              <button
+                class="flex items-center gap-3 text-gray-1"
+                on:click={() => removeRelay(relay)}>
+                <i class="fa fa-right-from-bracket" /> Leave
+              </button>
+            {/if}
+          </div>
+        </RelayCard>
       {/each}
     </div>
   {/if}
@@ -64,10 +82,16 @@
     <i slot="before" class="fa-solid fa-search" />
   </Input>
   {#each (search(q) || []).slice(0, 50) as relay (relay.url)}
-    <RelayCard {relay} />
+    <RelayCard {relay}>
+      <div slot="actions">
+        <button class="flex items-center gap-3 text-gray-1" on:click={() => addRelay(relay)}>
+          <i class="fa fa-right-to-bracket" /> Join
+        </button>
+      </div>
+    </RelayCard>
   {/each}
   <small class="text-center">
-    Showing {Math.min($knownRelays.length - $userRelays.length, 50)}
-    of {$knownRelays.length - $userRelays.length} known relays
+    Showing {Math.min($known.length - relays.length, 50)}
+    of {$known.length - relays.length} known relays
   </small>
 </Content>